back-end/requirements.txt
alembic==1.4.2
click==7.1.2
Flask==1.1.2
Flask-Cors==3.0.8
Flask-HTTPAuth==4.1.0
Flask-Migrate==2.5.3
flask-sqlacodegen==1.1.7
Flask-SQLAlchemy==2.4.4
inflect==4.1.0
itsdangerous==1.1.0
Jinja2==2.11.2
Mako==1.1.3
MarkupSafe==1.1.1
mysqlclient==2.0.1
PyMySQL==0.10.0
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
six==1.15.0
SQLAlchemy==1.3.19
Werkzeug==1.0.1
